refactor(example): split lab UI components
