+++ .github/workflows/npm-publish.yml
@@ -18,7 +18,8 @@ jobs:
         with:
           node-version: '24'
           registry-url: https://registry.npmjs.org/
+          package-manager-cache: false
       - run: npm install -g npm@latest
       - run: npm install
       - run: npm run prepack
-      - run: npm publish --provenance
+      - run: npm publish
